requirements.txt
flask
python-dotenv
flask-restful
flask_sqlalchemy
flask_marshmallow
sqlalchemy-filters
Faker